Dialer: Control dialer's incoming call proximity sensor check via an overlay
This feature can't be used on devices with proximity sensors that don't
work when the screen is off, e.g. sensors using ultrasound technology.
The result in those cases is a black screen with just the status bar
showing, instead of the normal incoming call UI.

Make it possible to opt-out via an overlay.

Change-Id: Ic3848d09e1ed80e5409cbecbaca2517db16ed0b8
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>